netfilter: nf_tables: fix nft_supported_family() parameter type
The caller of nft_supported_family() passes an integer to the function,
but the type of the parameter is defined as u8.  The mistake has made
while backporting.  Fix it to use same type.

Fixes: 24f1cf2efb69d ("netfilter: nf_tables: Reject tables of unsupported family")

+++ net/netfilter/nf_tables_api.c
@@ -746,7 +746,7 @@ err:
 	return ret;
 }
 
-static bool nft_supported_family(u8 family)
+static bool nft_supported_family(int family)
 {
 	return false
 #ifdef CONFIG_NF_TABLES_INET
